Move fetching of MetadataConfig to api
